Some cleanup of the util directory. Made improvements to the configureEnv.sh script... error checking etc.

+++ util/configureEnv.sh
@@ -10,6 +10,14 @@ downloadPY()
     mkdir -p "$BF_PATH_l"/externLib
     cd "$BF_PATH_l"/externLib
     
+    pyVers=$(python -c 'import sys; print(".".join(map(str, sys.version_info[:3])))')
+    if [ $(cut -d'.' -f1,2 <<< $pyVers) != "2.7" ]; then 
+        echo "ERROR: Need Python 2.7. Current version is: $pyVers" >&2
+        echo "Configure your environment such that the \"python\" command calls Python 2.7." >&2
+        echo "$pyVers"
+        return 1
+    fi
+
     # Create the virtual environment
     rm -rf "$virtEnvName"
     echo "Generating the virtual environment..."
@@ -17,9 +25,6 @@ downloadPY()
     retVal_l=$?
     if [ $retVal_l -ne 0 ]; then
         echo "Failed to establish the virtual environment." >&2
-        echo "Need Python 2.7. Current version is: " >&2
-        python --version
-        echo "Configure your environment such that the \"python\" command calls Python 2.7." >&2
         return $retVal_l
     fi
     echo "Virtual environment created."
@@ -114,7 +119,7 @@ downloadPY()
     echo
 
     echo "Copying BFfile.py to site-packages..."
-    cp "$BF_PATH_l"/util/basicFusion.py "$BF_PATH_l"/externLib/BFpyEnv/lib/python2.7/site-packages
+    cp "$BF_PATH_l"/util/BFpy/basicFusion.py "$BF_PATH_l"/externLib/BFpyEnv/lib/python2.7/site-packages
     echo "Done."
     echo
 
@@ -125,7 +130,11 @@ source \"$BF_PATH_l\"/externLib/BFpyEnv/bin/activate
     echo "$script" > ${HOME}/bin/activateBF
     chmod +x ${HOME}/bin/activateBF
     echo "Done."
-    echo "NOTE: Activate the virtual environment at any time by typing: source activateBF"
+    echo
+    echo "==================================================================================="
+    echo "= NOTE: Activate the virtual environment at any time by typing: source activateBF ="
+    echo "==================================================================================="
+    echo
 
     deactivate
 
@@ -144,6 +153,7 @@ downloadSched()
     local retVal_l
     local SCHED_URL_l="$1"
     local SCHED_PATH_l="$2"
+    local hn_l=$(hostname)
     # Change to the required PrgEnv module
     oldPRGENV=$(module list | grep PrgEnv | cut -f3 -d" ")
     if [ ${#oldPRGENV} -ne 0 ]; then
@@ -159,7 +169,7 @@ downloadSched()
 
     else
         retVal_l=0
-        if [ $hn != "cg-gpu01" ]; then
+        if [[ $hn_l == "h2ologin"* ]]; then
             module load PrgEnv-gnu
             retVal_l=$?
         fi
@@ -205,6 +215,7 @@ downloadSched()
 if [ $# -ne 1 ]; then
     echo 
     printf "\nUSAGE: $0 [-s | -p | -a]\n" >&2
+    pyVers=$(python -c 'import sys; print(".".join(map(str, sys.version_info[:3])))')
     description="
 DESCRIPTION:
 
@@ -217,7 +228,7 @@ ARGUMENTS:
 
 REQUIREMENTS:
 \tThe following requirements must be met:
-\t\t1. A Python 2.7 interpreter is visible. Current detected version is: $(python --version)
+\t\t1. A Python 2.7 interpreter is visible. Current detected version is: $pyVers
 
 "
  
@@ -264,6 +275,8 @@ SCHED_PATH=$BF_PATH/externLib
 
 DOWNLOAD_SCHED=0
 DOWNLOAD_PY=0
+OLD_CC=''
+OLD_CXX=''
 
 # Determine the command line options
 if [ "$DOWNLOAD_OPTS" == "-s" ]; then
@@ -280,18 +293,34 @@ fi
 
 # Detect whether we are on Blue Waters or ROGER, then load the modules that we need
 hn=$(hostname)
+ON_ROGER=0
+ON_BW=0
 
 # ON BLUE WATERS (we are assuming)
-if [ $hn != "cg-gpu01" ]; then
+echo "Checking which system we're on..."
+if [[ $hn == "h2ologin"* ]]; then
     PrgEnv=$(module list | grep PrgEnv)
     if [ ${#PrgEnv} -eq 0 ]; then
         module load PrgEnv-cray
     fi
     module load cray-mpich
-
+    # Blue waters, we need to use cc and CC compilers for C and C++ respectively...
+    OLD_CC=$CC
+    OLD_CXX=$CXX
+    export CC="cc"
+    export CXX="CC"
+    ON_BW=1
+    echo "DETECTED BLUE WATERS."
+    echo
 # ON ROGER
-else
+elif [[ $hn == "cg-gpu01" ]]; then
     module load mpich
+    ON_ROGER=1
+    echo "DETECTED ROGER."
+    echo
+else
+    echo "ERROR: Unable to detect which system we're on."
+    exit 1
 fi
 
 if [ $DOWNLOAD_PY -eq 1 ]; then
@@ -300,13 +329,25 @@ if [ $DOWNLOAD_PY -eq 1 ]; then
 
     if [ $retVal -ne 0 ]; then
         echo "Failed to setup Python." >&2
+        export CC=$OLD_CC
+        export CXX=$OLD_CXX
         exit $retVal
     fi
 fi
 
 if [ $DOWNLOAD_SCHED -eq 1 ]; then
     downloadSched "$SCHED_URL" "$SCHED_PATH"
+    retVal=$?
+
+    if [ $retVal -ne 0 ]; then
+        echo "Failed to setup NCSA Scheduler." >&2
+        export CC=$OLD_CC
+        export CXX=$OLD_CXX
+        exit $retVal
+    fi
+
 fi
 
-
+export CC=$OLD_CC
+export CXX=$OLD_CXX
 exit 0
